Fix core CHIP-8 implementation bugs
- Fix getMemoryAt parameter type from uint8_t to uint16_t for proper address handling
- Fix stack pointer boundary check (>= STACK_SIZE to > STACK_SIZE)
- Add error clearing on successful operations to prevent stale error states
- Improve error message formatting with hex addresses
- Add PC advancement on subroutine call errors to prevent infinite loops

+++ src/chip8.h
@@ -55,7 +55,7 @@ public:
   void setDrawFlag(bool condition);
   void setIndexRegister(std::uint16_t value);
   // Getters
-  std::uint8_t getMemoryAt(std::uint8_t address) const;
+  std::uint8_t getMemoryAt(std::uint16_t address) const;
   std::uint16_t getIndexRegister() const;
   std::uint16_t getProgramCounter() const;
   std::uint16_t getStackAt(std::uint8_t subroutine) const;
